requirements.txt
beautifulsoup4==4.12.3
bokeh==3.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.6.1
et-xmlfile==1.1.0
fonttools==4.50.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mysql-connector-python==8.3.0
networkx==3.2.1
ntlm-auth==1.5.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
plotly==5.20.0
prettytable==3.10.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.22
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
pyspnego==0.10.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-ntlm==1.2.0
requests-toolbelt==1.0.0
seaborn==0.13.2
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
sspilib==0.1.0
tenacity==8.2.3
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
unicorn==2.0.1.post1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
xyzservices==2023.10.1
